requirements.txt
fastapi==0.103.2
python-multipart==0.0.6
uvicorn==0.23.2
jinja2==3.1.2

PyPDF2==3.0.1
aiofiles==23.2.1

docx2python==2.6.3

ebooklib==0.18.1
pdfminer.six==20221105
python-docx==1.0.1

# New dependencies
PyMuPDF==1.23.5
Pillow==10.0.1
